requirements.txt
torch==2.9.1
torchvision==0.24.1
numpy
opencv-python-headless
OpenEXR>=3.4
timm==1.0.24
av
diffusers
easydict
einops
imageio
huggingface_hub
peft
Pillow
PIMS
setuptools
tqdm
accelerate
transformers
matplotlib
nvidia-ml-py
PySide6>=6.7.0
triton-windows>=3.5,<3.6; sys_platform == 'win32'
